requirements.txt
aiohttp==3.6.2
aioredis==1.3.1
amqp==2.5.2
asgiref==3.3.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
daphne==3.0.1
Django==3.1.4
django-cors-headers==3.2.1
django-filter==2.4.0
django-storages==1.9.1
djangorestframework==3.11.0
grpcio==1.34.0
grpcio-tools==1.34.0
hiredis==1.1.0
psycopg2-binary==2.8.6
python-decouple==3.3
requests==2.25.1
redis==3.3.11
supervisor==4.2.1
Twisted==20.3.0
txaio==20.4.1
whitenoise==5.1.0
channels==3.0.3
channels-redis==3.2.0
